panels are still better - will work out how to deal with the the hardcoded 70px later
